fix: isolate participant response contracts
