+++ web-next/lib/api/admin.ts
@@ -14,6 +14,12 @@ import {
   waitlistRejectResponseSchema,
   type WaitlistRequest,
 } from './waitlist-contract';
+import {
+  participantAddResponseSchema,
+  participantListResponseSchema,
+  participantRemoveResponseSchema,
+  type Participant,
+} from './participant-contract';
 
 export type { GovernanceWeights, ContentRules };
 export type { WaitlistRequest } from './waitlist-contract';
@@ -22,35 +28,12 @@ export {
   waitlistListResponseSchema,
   waitlistRejectResponseSchema,
 } from './waitlist-contract';
-
-const participantDidSchema = z.string().startsWith('did:').min(5);
-const participantTimestampSchema = z.string().datetime({ offset: true });
-
-export const participantSchema = z.object({
-  did: participantDidSchema,
-  handle: z.string().min(1).nullable(),
-  added_by: z.string().min(1),
-  notes: z.string().nullable(),
-  added_at: participantTimestampSchema,
-}).strict();
-
-export const participantListResponseSchema = z.object({
-  participants: z.array(participantSchema),
-  total: z.number().int().nonnegative(),
-}).strict();
-
-export const participantAddResponseSchema = z.object({
-  success: z.boolean(),
-  participant: z.object({
-    did: participantDidSchema,
-    handle: z.string().min(1).nullable(),
-    notes: z.string().nullable().optional().transform((value) => value ?? null),
-  }).strict(),
-}).strict();
-
-export const participantRemoveResponseSchema = z.object({
-  success: z.boolean(),
-}).strict();
+export {
+  participantAddResponseSchema,
+  participantListResponseSchema,
+  participantRemoveResponseSchema,
+  participantSchema,
+} from './participant-contract';
 
 export interface RoundSummary {
   id: number;
@@ -112,8 +95,6 @@ export interface RoundDetails {
     created_at: string;
   }>;
 }
-
-export type Participant = z.infer<typeof participantSchema>;
 
 export interface AdminStatus {
   isAdmin: boolean;
